# docs/openapi/auth.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RunAsh Auth API",
    "version": "1.0.0",
    "description": "OpenAPI contract for core auth routes and enabled auth plugins."
  },
  "servers": [
    {
      "url": "/",
      "description": "RunAsh deployment"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session-token",
        "description": "Better Auth session cookie."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "scimBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "SCIM provisioning token."
      }
    }
  },
  "paths": {
...
    "/api/auth/anonymous": {
      "post": {
        "operationId": "post_api_auth_anonymous",
        "tags": [
          "Core Auth"
        ],
        "summary": "POST /api/auth/anonymous",
        "security": [],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/auth/forgot-password": {
      "post": {
        "operationId": "post_api_auth_forgot_password",
        "tags": [
          "Core Auth"
        ],
        "summary": "POST /api/auth/forgot-password",
        "security": [],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/auth/siwe/verify": {
      "post": {
        "operationId": "post_api_auth_siwe_verify",
        "tags": [
          "SIWE"
        ],
        "summary": "Verify Sign-In with Ethereum",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "examples": {
                "siweVerification": {
                  "value": {
                    "nonce": "nonce_123",
                    "message": "runash.ai wants you to sign in with your Ethereum account...",
                    "signature": "0xabc123",
                    "walletAddress": "0x71C7656EC7ab88b098defB751B7401B5f6d8976F",
                    "chainId": 1
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}